JAVA-17: Move Running Setup Data on Startup in Spring into spring-boot-data
