+++ spring-boot-modules/spring-boot-data/README.md
@@ -10,3 +10,4 @@ This module contains articles about Spring Boot with Spring Data
 - [Repositories with Multiple Spring Data Modules](https://www.baeldung.com/spring-multiple-data-modules)
 - [Spring Custom Property Editor](https://www.baeldung.com/spring-mvc-custom-property-editor)
 - [Using @JsonComponent in Spring Boot](https://www.baeldung.com/spring-boot-jsoncomponent)
+- [Running Setup Data on Startup in Spring](https://www.baeldung.com/running-setup-logic-on-startup-in-spring)
