+++ spring-boot-modules/spring-boot-data/src/main/java/com/baeldung/startup/LogicInConstructorExampleBean.java
@@ -1,26 +1,22 @@
 package com.baeldung.startup;
 
-import java.util.Arrays;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.InitializingBean;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Scope;
 import org.springframework.core.env.Environment;
 import org.springframework.stereotype.Component;
 
+import java.util.Arrays;
+
 @Component
 @Scope(value = "prototype")
-public class InitializingBeanExampleBean implements InitializingBean {
+public class LogicInConstructorExampleBean {
 
-    private static final Logger LOG = LoggerFactory.getLogger(InitializingBeanExampleBean.class);
+    private static final Logger LOG = LoggerFactory.getLogger(LogicInConstructorExampleBean.class);
 
     @Autowired
-    private Environment environment;
-
-    @Override
-    public void afterPropertiesSet() throws Exception {
+    public LogicInConstructorExampleBean(Environment environment) {
         LOG.info("Env Default Profiles", Arrays.asList(environment.getDefaultProfiles()));
     }
 }
